code clean up for improved bot
